chore: move common test functions to support
