+++ src/backend/integration/src/support/index.ts
@@ -1,3 +1,4 @@
+export * from './common';
 export * from './date';
 export * from './governance';
 export * from './identity';
